16bpcgenのC++03化（微修正）

+++ c++/16bpcgen/v2/src/PixelConverters.cpp
@@ -79,7 +79,7 @@ Pixel<uint16_t>& Reversal::convert(Pixel<uint16_t>& pixel)const
 Gamma::Gamma(const std::vector<Pixel<uint16_t>::value_type>& lut, Ch ch):
 	Channel(ch), lut_(lut)
 {
-	if(lut_.size() != Pixel<uint16_t>::max + 1){
+	if(lut_.size() != static_cast<std::size_t>(Pixel<uint16_t>::max + 1)){
 		throw std::runtime_error(__func__);
 	}
 }
